requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.2.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.0.1
flask_bcrypt==1.0.1
flask_session==0.5.0
Flask-Uploads==0.2.1
email_validator==1.3.1
python-dotenv==1.0.1
flask_bcrypt==1.0.1
xlsxwriter
pyodbc>=5.0.0
pandas
gunicorn
python-dateutil
pymysql
weasyprint